Report progress correctly when amount of data is more than 2GB
2013-01-16  Jon TURNEY  <jon.turney@dronecode.org.uk>

	* threebar.cc (ThreeBarProgressPage::SetBar2): Change from long int to
	long long int to handle more than 2GB of data.
	* threebar.h (ThreeBarProgressPage::SetBar2): Update prototype.
	* install.cc: Change md5sum and install counters to long long int.
	* geturl.cc: Change download counter to long long int.
	* geturl.h: Ditto.

+++ geturl.cc
@@ -55,8 +55,8 @@ extern ThreeBarProgressPage Progress;
 static int max_bytes = 0;
 static int is_local_install = 0;
 
-int total_download_bytes = 0;
-int total_download_bytes_sofar = 0;
+long long int total_download_bytes = 0;
+long long int total_download_bytes_sofar = 0;
 
 static DWORD start_tics;
